modifications for replays
use-gps-as-height-on-ekf-to-init (#876)
use-gps-data-with-increased-precision
never-turn-off-yaw-fusion (#905)
do-not-use-mag_heading_noise-as-gps-yaw-variance
EKF-allow-ekf-init-without-magnetometer
control-fix-to-startGpsHgtFusion-again-after-timeout (#951)
allow-EKFGSF_yaw-to-work (#902)
remove-default-15-m-s-airspeed

+++ test/sensor_simulator/gps.h
@@ -54,9 +54,9 @@ public:
 	void stepHeightByMeters(float hgt_change);
 	void stepHorizontalPositionByMeters(Vector2f hpos_change);
 	void setPositionRateNED(const Vector3f& rate) { _gps_pos_rate = rate; }
-	void setAltitude(int32_t alt);
-	void setLatitude(int32_t lat);
-	void setLongitude(int32_t lon);
+	void setAltitude(float alt);
+	void setLatitude(double lat);
+	void setLongitude(double lon);
 	void setVelocity(const Vector3f& vel);
 	void setYaw(float yaw);
 	void setYawOffset(float yaw);
